renamed grug_value.py to types.py and moved grug type declarations to types.py
